:recycle: Split endpoints into multiple class
