+++ classifai-core/src/main/java/ai/classifai/util/Hash.java
@@ -18,10 +18,9 @@ package ai.classifai.util;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileInputStream;
+import java.io.*;
 import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
 
 /**
@@ -32,36 +31,25 @@ import java.util.Base64;
 @Slf4j
 public class Hash
 {
-    public static String getHash256String(@NonNull File filePath)
-    {
-        try
+    public static String getHash256String(@NonNull File filePath) throws IOException, NoSuchAlgorithmException {
+        byte[] buffer= new byte[8192];
+
+        int count;
+
+        //not thread safe, hence single instance per function
+        MessageDigest digest = MessageDigest.getInstance("SHA-256");
+
+        BufferedInputStream bis = new BufferedInputStream(new FileInputStream(filePath));
+
+        while ((count = bis.read(buffer)) > 0)
         {
-            byte[] buffer= new byte[8192];
-
-            int count;
-
-            //not thread safe, hence single instance per function
-            MessageDigest digest = MessageDigest.getInstance("SHA-256");
-
-            BufferedInputStream bis = new BufferedInputStream(new FileInputStream(filePath));
-
-            while ((count = bis.read(buffer)) > 0)
-            {
-                digest.update(buffer, 0, count);
-            }
-
-            bis.close();
-
-            byte[] hash = digest.digest();
-
-            return Base64.getEncoder().encodeToString(hash);
-
-        }
-        catch(Exception e)
-        {
-            log.info("Error in generating SHA-256 Hashing: ", e);
+            digest.update(buffer, 0, count);
         }
 
-        return null;
+        bis.close();
+
+        byte[] hash = digest.digest();
+
+        return Base64.getEncoder().encodeToString(hash);
     }
 }
